Implemented create method and "privatized" constructors
The picamera2 library seems to throw RuntimeError if anything goes wrong, so wrapping the camera creation in a try-except block.

Picamera2 fails capture_array() the same way opencv would for read()

Changed the configuration for the picamera2 camera to use a still configuration. TL;DR most configurations are similar in parameters, except for buffer size. See manual/source for more